DialogWidget improvements
